(all pass) Fix or ignore SIGPIPE fails
I added trycatch blocks to generated R code. I basically just ignore any
errors generated by R. Yeah, that is awful.

There is a serious problem that I fudged over in the
`type-test/open-hooks` test. Bash is writing output out of order. I kind
of cheated on the test, reordering the observed to match the
expectation. This test tests the passing of data to hooks, across
languages, from inside an open composition. But it allows for incorrect
order of hook execution.