Add naive two language example
This is an example of a simple two piece analysis. One part gets data
from online, unzips it, and stores the resulting file. The second part
reads the file and performs some statistical analysis.

I really need tables. There is a lot of boilerplate in this example
which I could eliminate if I had a table type with good conversion
handling.