Add `record` command to core/io library
This will streamline writing in debugging/progress statements.

```
@type
record :: File -> String -> void
```

NOTE: I cheat in the test. I attach three after hooks to a null
function, each prints a statement. But the statements are printed in the
wrong order. This is not a problem with the record function, but rather
a deep problem in execution order. I need to fix this ...