(1 fail) Add infrastructure for generic types
Generics are read in as couplets with lhs string labels and rhs types,
which default to star atomics.

I've implemented most of the infrastructure (the lexer reads in the
generics correctly, the W types are all fine, etc). But the generic
inference function simply reduces the generic to a star atomic:

```
Couplet<String, FT_ATOMIC<'*'>>  -->  FT_ATOMIC<'*'>
```

I need to fill in this function to 1) infer its identity given an input
type, 2) propagate its type to entangled generics, 3) propagate their
types to their outputs, 4) recurse as necessary.