Require generic symbols are single characters
Generic types in Haskell are required to start with a lowercase letter
and they conventionally are single characters. This is reasonable for a
generic; longer names belie the polymorphism of a generic.

Previously LOC generics had to be lowercase but could be any length. Now
the single character convention is a syntactic requirement.

I may relax this requirement later, but it simplifies type inference,
allowing direct mapping from symbol name to a constant length vector.