(8/74 fail) Add type compatibility function
I am working on implementing a basic Hindley-Milner algorithm for type
inferrence. One necessary component is a function to determine if two
types have the same form.

Let a and b be nodes in a type tree.
Where
  Node :: Generic | Primitive | Constructor [Node]
The Constructor nodes have 1 or more Node children

a and b are compatible if they are specifications of a common Node

comp a b =
  | Generic _ = TRUE
  | _ Generic = TRUE
  | Primitive Primitive = a == b
  | Constructor Constructor = class(a) == class(b) AND all (map (zip a b))
  | otherwise = FALSE

I really wish I were writing this in Haskell ...