Found a bug in previous commit
The current implementation of `all_io_types_are_compatible` correctly
handles differing forms of constructors, for example

```
a     > (a,b,c)
(a,b) > (a,(b,c))
(a,a) ! (a,a,a)
```

Where `>` means more general and `!` means incompatible. For two types
to be compatible it is necessary that every node in their type trees
either be the same or be a specialization of the other. However, this is
not a sufficient rule. Generic type entanglement also needs to be
preserved. For example

```
(a,a) ! (a,[a])
```

Because there is no substitution of a such that both `a<=>a'` and
`a<=>[a']` are true.

So I need to keep track of which substitutions I must perform on each
side in order for the two to be equal. The all substitutions must be
isomorphic.