Simple generic inferrence now works
Generics seem to pass and propagate correctly, although much more
testing is needed.

The types of generics inside containers can be inferred, however, these
wrapped generics cannot be used to infer other generics. For example,

```
@type
foo :: a -> b -> (a, a)
bar :: (a, c) -> a

@path
bad . foo . 1 2
```

The output of `foo` is correctly inferred as `(Int,Int)`, but the input
to bar remains as `(*,*)`.