Update handling of backticks in arguments
I previously passed these as positionals, but that isn't really what
they are in the argument context. Rather, backticks imply exact passing
of the material to the target language. For example:

```
foo=`runif(100)`
```

Would be a perfectly valid argument in R, where `foo` is given the
default value of a vector of 100 uniform random variables.

As for passing positional arguments, I don't need any special handling,
if there is no `=` sign, then they are positional.