Convert get-cds to extract-fasta
Rather than converting to CDS, which requires a map operation and
needlessly complex steps, I will just present a Fasta extractor that
loads a GFF, passes it through a filter, renames it, and then extracts
the fasta sequence. This pipeline is complex enough to showcase the
advantages of Morloc but small enough to fit in less than 50 lines.

Here are a few thoughts on types. Currently I am expressing GFF with the
overly general `Table` type. To load the thing into R, I have need a
dedicated function that sets the column names and types. I shouldn't
have to handle IO, the compiler should figure that out from the type.
However, my type system is currently not expressive enough. I want to be
able to do the following:

```
readG ::
  File ->
  Table {
    seqid:String,
    source:String,
    name:String,
    start:Int,
    stop:Int,
    score:Num,
    strand:Char,
    phase:Int,
    attr:String
  }
```

Or even better:

```
typedef GFF ::
  Table {
    seqid:String,
    source:String,
    name:String,
    start:Int,
    stop:Int,
    score:Num,
    strand:Char,
    attr:String
  }
readG :: File -> GFF
``

Or perhaps I can define types in the ontology section

```
@ontology
GFF ::
  Table {
    seqid:String,
    source:String,
    name:String,
    start:Int,
    stop:Int,
    score:Num,
    strand:Char,
    attr:String
  }
```

I also need to be able to specify relationships between types, for example:

```
GFF --> HTable --> Table
```

Indicates that a GFF is convertable to a headered table, which in turn in
convertable to a headerless table. Similarly

```
[a] --> Table  # a one-column table
Natural --> Int --> Num --> Complex
```

Some types are always interconvertable

```
String <-> Text
```

The only difference between these two is how they are passed to as arguments,
one as a String, one as a file. But maybe this distinction is more general.
Maybe I need a deeper syntax for it. Something that indicates how compact
something should be. Say String<C:1> and String<C:2> where C:1 indicates low
compression and C:2 higher compression. I'm not sure of the syntax. I should
read up on ithkuil ... there are probably many interesting ideas that I could
borrow from there.