Add sample JSON files for all major types
The common data representation will be JSON. JSON is simple, easy to
read, memory efficient, and rich enough for most cases.

Eventually I may need to replace JSON with the more general YAML format.
The most important thing that YAML can do that JSON can't is use
pointers. This is required to efficiently express certain data
structures. JSON is a strict subset of YAML, so I can switch to YAML in
the future without breaking existing code.

For now, the format of the common form will be:

```
{ "type":<TYPE>, "value":<VALUE> }
```

I may add more later, but I think additional fields should be optional.
I could remove the `type` field, but for now it is beneficial since it
allows the data to stand alone. Also it is useful for debugging and
internal validation of state. Having the data describe itself also
increase modularity. For example, the `convertE` function will change
the `value` and the `type`, then anything downstream can know the type
from the JSON file, without having to learn from outside that there was
a conversion. It can also be used to propagate error, where the error
form may be:

```
{ "type":"ERROR", "value":"This is an error message" }
```