Abstract out the "mode" of interpretation
A given morloc program might be interpreted in more than one way. For
example, you might just want to type check it, or convert it to an
intermediate format, or do a dry run, of do the full thing.

Here is the current flow (abstracting away a few details):

```
eval mode . parser . lexer $ code
```

All modes are defined in `Morloc.Mode`, currently this only contains
`asLIL`.

The Morloc::interpret function now has the signature

```
Mode -> MorlocCode -> Either ErrorMsg Output
```

The Main module is responsible for choosing the mode and loading the code
(as a string); passing mode and code to `interpret`; and printing the
result.