Refactor modes to make them binders for IO
Previously the "modes" were functions that took the AST from eval and
produced a String of output. This works from some things, like
generating LIL or showing the generated code. But in general, the output
of Morloc is effectual. I need to be able to show the final output,
which comes from making calls to the manifold pools. Anyway, I've now
reimplemented them with to output `IO ()` instead of `String`.

Next I will work on adding the mode `asResult`, which will print the
final output of a Morloc workflow.