Major refactor
 * remove outdated comments (including the derivations)
 * removed LIL generation (I don't need it anymore)
 * remove NodeAttribute.hs
 * add Data.hs and MData
 * replace setid function with a clean zip function, that lets me zip a
   series (as defined by a successor function and initial element) into a
   graph.

MData is the biggest addition. Every type that Morloc supports should be
expressed in this construct. The Parser directly writes MData into the
AST, the evaluator creates (Graph MData) graphs, the generatory zips
across the graphs to generate the code. This is much more cohesive than
having a Syntax, Type, and NodeAttribute, all which stored the same
information, in subtly different ways.

Next I will add support for homogenous lists of primitives.