Slightly update common types
Determing common representations for the types will require a lot of
careful thought. There are an infinite number of possible types (tuples,
for instance nest to arbitrary depth). Ultimately, we will have to
generate the code needed to parse the types based off the type string.
The specific languages should have functions for parsing elemental
types. The Morloc core will need to be able to combine these elements to
represent more complex types. I have a tingling sensation that this is
deeply related to parsing, but the reverse in some sense. Sparkly magic
from the combinator wizards should help us out here.

We have three things: 1) a type string, 2) data in the common form, 3)
data in a language X structure. We need a parser that takes the #1 and
generates two parsers in language X: one parser translates #2 into #3,
the other does the reverse. The parsers in language X are built from a
collection of simple parsers in language X.