Fix #3 - remove hard-coded shebang directive
This is not the final solution. Previously, the generated files
specified their own interpreters through the shebang directives. Now the
calling script directly determines the interpreter with direct calls to
`bash` or `Rscript`. In this way, we use whatever executable is in PATH.
This works fine if the programs are installed and have the appropriate
names. However, there are cases where this will fail. `python`, for
instance, may refer to python2 or python3.

What I need is a way to specify these paths in a config file. More
generally, I need a way to specify dependency requirements.