Replace REPL with file parser
Currently this file parser just reads one file, feeds it line by line
into the Morloc interpreter, and shows the parse of each line.

```
sum [1,2,3] ==> Right (Node (MFunc "sum") [Node (MInts [1,2,4,8]) []])
1           ==> Right (Node (MInt 1) [])
1 1         ==> Left "BadApplication: Primitives cannot take arguments\n"
```

This is nice for debugging, but this isn't what we ultimately want.

I need to settle on a syntax. It should be as minimal as it can be. None
of the fanciness of the old C Morloc. Maybe just good, old fashioned,
function application.