Last commit before merging in smurf code
I developed a new parser with a nearly complete morloc syntax in the
temporary repository @arendsee/smurf. I will merge that code in. This
will involve a complete replacement of the current Lexer.hs, Parser.hs,
and Data.hs files. Then a complete rewrite of the evaluator. This
process will be spread out over several commits, many of which may be
unstable.