Evaluate the script into a Program object
Nothing in `Program` is quite is what it is said to be yet.

 * The workflow isn't really the workflow, but rather a list of
 functions.  Each function may link to values inside other functions. It
 is these linkes that create the workflow.

 * The ontology this isn't really the ontology, but rather just a list
 of type signatures. The ontology will hold the relations between them.

 * The packages aren't really packages, just the in-script imported
 code, with none of the metadata, export lists, and other info that a
 real package should have.

I'll need to start banging these into what they are suppose to be. But
before I do that, I want to take a very simple subproblem and implement
it all the way through the code generation stage.

Given a script of independent declaration/composition pairs, and a
collection of source code from known languages, build a working CLI
program.