Add CodeGenerator language template
My rough idea is to have a record that contains all the functions needed
to generate a script for a specific language. To add support for a new
language, the user will have to define all the functions in the record.

I also need to give more thought to the import/export system. All
functions ultimately come from those defined in `source <lang>` bodies.