BROKEN - made rough generatePools function
It doesn't type check yet. The problem is somewhere in the giant
composition in `makePooler`. I'll fix this next.

After it is doing what it is meant to, the next step will be to
implement the `generatePool` function:

```
generatePool :: [FunctionTree SNode] -> Source -> Pool
```

The `makePooler` function feeds `generatePool` all the info it needs to
build the pools (for now at least). The `FunctionTree Snode` structures
contain each node, its children, and their sources.