Now function bound arguments are threaded through
My code is getting to be really ugly. I'm going to need to refactor the
hell out of it. Also, much of what I have is specific to R. It will need
to be generalized a lot.

Anyway, next thing up is to add the appropriate BEGIN and END materia to
the pools. This should allow functions to be called with arguments. The
arguments need to be parsed into R data. For now, I can just support
strings and numbers, so this should be pretty easy.