And simple R is now online!
Now I need to get the manifold nexus cracking. Need a little
documentation and the ability to call specific function by name.

Once this is up, I will be able to at least demo R code.

After that, I can work on adding python. This will require a bit of
generalization of my generation code (not much, I don't think).

Once python is running, I will need to get the foreign call
functionallity up and running. This should not be too hard in the most
naive case.

Add a few bells and whistles, and I will be back to where the C Morloc
was.

After that, I will need to start working on the type system. Simple
typechecking, then constraints, then using the type system to make
smarter foreign interfaces.