Fix application of arguments
There was a problem in my parser, where expressions like "f a b" where
read as: `f (a (b))`. That is `a b` was read in as an application. I've
fixed this. But I am not happy with my solution. The data structures I
am using seem increasingly ad hoc and messy. I need a serious refactor,
maybe applying the magic of foresight this time.