Stub implementation of triple parser
Now the parser returns a `[Triple]` instead of a `[Top]`. Everything is
parsed, but internally the results are ignored and stub Triples are
returned. This completely bypasses the Evaluator and Generator.

Also, I also removed all Triple handling and explicit state dependence
from the Lexer. I need to maintain the purity of the Lexer, keep it
simple and wherever possible extend its role. Ultimately, it would be
nice to make it a "real", context-free lexer that returns tokens and
then let the parser by a context-sensative token-stream parser that
builds the RDF store. Then build the type-checker, type-casting, and
generators on top of the RDF store. I find myself gradually working my
way back to the C roots, where I used the clean Loc Intermediate
Language (LIL).