Add a few more tripples
Progress is slow but steady. The use of literal strings for relations
and subjects worries me. The wild freedom it offers comes at a high
cost. I perhaps should codify the legal relations in an enumeration.
Also choices in my RDF design, the knowledge modeling, is being driven
by ease of parsing. I need to think it through more carefully. Though
perhaps this is not so important within this iteration. This is not the
final rewrite.