Refactor and detect a flaw
The "scope" I record in the state is not really scope at all. It is just
the parent. This isn't enough. Now the tuple `((A,B),C)` complies into
the RDF:

```
2   :isa          :tuple
2   :has_parent   1
3   :isa          :tuple
3   :has_parent   2
4   :name         A
4   :has_parent   3
5   :name         B
5   :has_parent   4
6   :name         C
6   :has_parent   5
```

B is properly recorded as the child of A. This allows unnested tuples,
such as (A,B,C), to be recorded with internal elements related as a
linked list. It breaks down though if there is any nesting. In the
example above, C, is treated as the direct child of B, rather than the
child of (A,B) as it should be (or should it be?).

Maybe I need a vertical and lateral concept: a `up` and `prev`.