Partial solution to test 2
I've updated the order of output in `declaration`, but the main reason
this test fails is in `application`. In `f x = x`, the single rhs
variable `x` is treated as an application (e.g. a function call with no
arguments). It really should instead be treated as a variable. Indeed,
the first element in any application should be treated as a variable
with following elements as arguments to the variable. A variable with
arguments is assumed to be function.