Type signature tests now pass
Changes
  * Removed nodes "boolExpr" and "term"
  * Added ":binop" object with ":lhs" and ":rhs"
  * Added ":input", ":output", and ":constraint" relations

The "boolExpr" node stored the fact that the expression was boolean
(e.g. a constraint). Essentially, it was storing the output type of the
binary operator. This is redundant, since all the operators will have
their type stored already (eventually). Ensuring they are the correct
type is the role of the typechecker.