add test for `foo :: X`
This test fails. The reason is that the current type signatures consider
everything a function (possibly with no output). If the function has no
output, an `(i, ":output", 0)` triple is printed. This makes no sense
and can be fixed easily. But the fact remains that the `X` above is
listed as an `:input`.

Should this be a function? How should a function look that takes no
argument, but returns a value? If `foo :: X` is a function, is `X` the
input or output? Since both are reasonable, `foo :: X` should NOT be a
function declaration but a type declaration. Then `foo :: () -> X` and
`foo :: X -> ()` would be the functions with no input or output,
respectively.