Add test and hanlding for HOF types
I need to carefully distinguish the difference between a string ":value"
and ":name".

":value" is used as a relation to describe what something contains.

":name" is used when the language object is an unquoted string.

We should not use ":name" in the place of ":value". I was translating
things like "Int" into

```
(i, ":isa", ":type")
(i, ":name", "Int")
```

But this is inconsistent. Because then other types like arrays or
records, would have ":value" instead. It is more consistent to let every
type have a ":value", so:

```
(i, ":isa", ":type")
(i, ":value", "Int")
```