Replace "child" with "contains" where proper
For tuples, lists, records, unary operators (+/-).

Type parameters are still listed as "child".

I have somewhat mixed feelings about this. On the one hand, more
descriptive relations are easier to read. But on the other, they are
harder to remember. I may go with something inbetween. Where the
relations are internally encoded in an enumeration that is very
specific. This enumeration will then be translated into RDF in one
place. This will stabilize the code and allow clean control over the RDF
format. Ultimately, the RDF needs to be extremely stable, since changes
to it may create backwards incompatibility with tools built on Morloc.