Add test for arithmetic functions in constraints
Now can call `f x y` in a constraint. I also made the relation of `f` to
`x` and `y` be ":argument".

These relationships do need precise names because their may be multiple
out nodes from any node. `f` might also have a link to a documentation
function. The doc node is also a child, in that it is downstream of `f`.
But we certainly do not want it to be confused with the arguments for
`f`.

I also need to sync these changes with the boolean functions. Since
those are (perhaps unwisely) parsed separately.