Complete initial test suite
I've completed the TODO list with tests for compilation to RDF most
features of the language.

Eventually I will crystallize the RDF implementation into algebraic
types Haskell-side and a formal specification RDF-side. But first I want
to get my code generator back on line.

To do this I will make a somewhat temporary library of functions for
operating on `[Triple]`. This will be hghly inefficient. But if I design
it well, I will be able to swap out the function bodies and the
`[Triple]` type later for performant alternatives without disrupting the
code generator and type checking logic.

Essentially, I need to make a functional API to the tiple store.

The emerging architecture looks something like this. The Morloc script
is compiled into RDF. This data can be stored, shared, or whatever. Next
I have an interface to the triple store that uses graph semantics to
extract data. The typechecker, type inference machine, and code
generator all use this interface. The triple store also passes through
at least two phases: 1) the raw output of the parser and 2) the
validated output of the type checker.