Merge rdf4h
I now parse everything directly into an in-memory RDF object. Also I can
write the program in Turtle format. I have deleted all the material
above the parser. Also all the tests are now broken.

TODO:
 [ ] fix the RDF tests
 [ ] write new evaluator that makes `[Manifold]` from the RDF graph
 [ ] write new typechecker
 [ ] write code generator

There are essentially 4 representations of the program.

 1. Morloc script and source code for all functions
 2. RDF
 3. Haskell internal structures `[Manifold]`
 4. Executable

The Morloc script will ultimately diversify into many flavors. But all
will compile to the same RDF (or at least an RDF with a common core).
The RDF is a general, portable description of the Morloc program.
Eventually I can also load all the required source code into the RDF.
The Haskell internal structures and the executable are all specific to
the implementation.