I am reviewing the situation
To this point, I've relied on the triples being ordered. It is generally
not safe to assume this of an RDF store. To fix this problem I will need
to number each element in each list.

Also in the tests, I will need to convert triples to an ordered list
(rdf4h has a function for this).