Do not quote literal strings
There is a lot of work left to do on this RDF. It may be time to sit
down and hammer out the formal specs. Or I can just keep hacking until
all my tests pass and then move on to the writing the evaluator. Yeah,
why don't I just hack a bit longer ...