Stub evaluator and data
The old Data.hs file is back, with some modifications.

I wonder though if any of it is really needed. Perhaps I could dispense
with all data structures and work directly with the RDF graph.

I am not very familiar with querying RDF through rdf4h. As I build the
evaluator, I will hone a set of commands that may make this easier.
Eventually, if I find it is so easy the dedicated `Manifold` record is
uneeded, great.