Let the parsec state generate UNode ids
Previously the IDs generated by state where a series of integers. Then I
converted these to IDs. But this gets awkward when I have to extract the
ID from an UNode, where it is stored as Text. Now I let state generate
the UNode IDs and use these everywhere, with no Int to Node shenanigans.