Add `fetchConstraints` and change constraint edge
Now there is a relation (URI, "morloc:constraint", URI). Previously I
had the constraints as indexed children of the type. However the type
could have other children (parameters), leading to ambiguities. Also,
logically a type is not a list of constraints. Finally, the constraints
are not ordered, so index semantics is not needed.

On another note, graph traversal under my new system is getting pretty
easy. The one annoying thing is having to pass rdf to every function. I
should wrap it into a Reader monad.