Add export statements
Now the statements that are exported need to be specified. For example

```
source "R" ("runif")
export runif
```

The alternative is for Morloc to guess what you want to export, and
there is no good way to do that.

Eventually I will replace the export keyword with cleaner module syntax.
The 'export' predicate will be one of several that describe a function.
But that is all down the road.