Add stub nexus call generator and import walkers
There is a knot I need to unravel here. Functions can be defined in
Morloc. The functions disappear upon compilation (only the functions in
the source languages are "real"). They are not just sugar, though, since
they do create scope.

When a Morloc function is exported, what pool should it be called from?
The could be called from any pool, but the best performance will obtain
if they are called from the pool that their child function is called
from.