Now collapse name into alias when no alias given
Previously when no import alias was given, I just created no entry for
alias. Then the downstream programmer had to infer from the absence of
an alias that the Morloc name for the function was the given name.

For example in

```
source "R" ("fooooo")         -- S1
source "R" ("fooooo" as foo)  -- S2
```

In S1, no alias triple would form. Now alias triples are created for
both. With the alias in S1 defaulting to "fooooo". This simplifies
things downstream and even in the parser itself.