Write some Haddock documentation
There is a lot of work to be done here in the documentation arena.

 * A header for every module
 * A description (with an example preferrably) for every function
 * A description of every field in every record (e.g. MorlocError)

I also need to update the README. There is a lot of very legacy material
in there. I have resisted removing it in the past since it is still more
advanced than what I have now. However the gap is closing (already I
allow arguments to functions, which was not possible before), so maybe
it is time to scrap the old.

Finally, I need to organize all the material in `docs`. Much of this is
highly speculative or dated work. I should organize it into a manual.
Possibly also I should merge it with the extensive additional material I
have in other locations.