Simplify compiler flags
Recently GHC converted from the `f*` to `W*` syntax for warning flags.
This is nice since it makes the GHC flags more in line with the
conventions of gcc. However, Travis uses an ancient version of Haskell.
I could convert back to the old forms. But instead I will just delete
the extra flags I was using, going back to just Wall.