Allow sourcing of foreign code
The next issue to deal with is typecasting the user-provided arguments.
What I ought to do here is feed the users arguments into a parser that
can convert them to Morloc types (Int, Float, String, List, Tuple, and
Record).  Or they could enter raw JSON for anything other than an Int,
Float or String.

I've partially solved this proble before, in the old version of Morloc.