Simplify packers
The previous commit overcomplicated the problem. The serialization
functions can have the forms:

```
<name> :: packs => <lang>/<type> -> Common
<name> :: unpacks => Common -> <lang>/<type>
```

In order to find the appropriate function, the compiler needs to have a
mapping from the language-specific type, `l a`, to the general type,
`b`.

I've also moved to the syntax:

```
{:
    <subject> <predicate> <object> .
    <subject> <predicate> <object> .
:}
```

For representing arbitrary relations. This is could benefit from some
Turtle-style sugar, but is fine for now.