Fix test code
Made it ugly as hell. Perhaps there is a way I can use the prefixes in
the strings and then expand them?

Or better, I could instead compare the turtle output of the parsed code
to expected files.

My tests are far too fragile right now. Keeping them in sync with the
code base is a burden.