Integrate uploading into a single build command
Now a Morloc program can be built from script to executable with the
single command:

```
morloc <script> <endpoint>
```

For example:

```
morloc sample.loc http://localhost:3030/morloc
```

I am not happy with the burden of setting up a SPARQL endpoint resting
on our users. I will find a way to circumvent this problem in the near
future. Ideally, I would like an embedded SPARQL database, something
like SQLite. It would need to be an open source tool, not some
proprietary trap. This means the options are a little limited.

It should also be possible to do all SPARQL ops through MorlocIO. In
this case, the compiler would build script into an RDF graph, upload it
to MorlocIO, and then query the Morloc.IO database to build the program.
There would be a performance penalty here, though perhaps it would be
small. And there would be oppurtunity for MorlocIO to perform extra
optimizations and check for updates.

However, MorlocIO should always be an optional component of the
ecosystem. Compiling should be possible locally without requiring any
connections. In the long term, remote access will become increasingly
expensive was we spread more than a few light-seconds apart.