Add Haskell interpretation to string interpolation
Now arbitrary Haskell code can be executed in quasiquotes.

This is a pretty normal thing to do. There are a couple other libraries
that also allow this. What sets my quasiquoting apart is that I am doing
it with text builders, not straight text or strings.

I ought to sync my syntax conventions with theirs. For example, using
`#{...}` for interpolation. I will not change yet, though, since it is
not clear yet to me how far I will diverge from standard practice. Also,
since Morloc is not just a Haskell project, and the templates should be
readable to the widest possible programming audience, I ought to look
into non-Haskell conventions.