Determine pool for Morloc functions
Previously the nexus generator didn't even include Morloc compositions.
Now it places the composition manifold in the pool of the language that
is first called.

The code that performs all this is ugly as hell.

Oh, I also changed all the `mlc:value` properties to `rdf:value`.

Next thing to do is extend the pool generator to handle the other
manifold forms.

I also need to sit down and think through what exactly all the manifold
forms are.

Then I need to write a bunch of tests.

Then refactor my code until it sparkles.

One thing that would tidy up my code is a serious templating that would
build tuples from the sparql return tables. For example:

```
foo :: SparqlEndPoint -> IO [(Text, Maybe Text)]
foo = [sparql|
    SELECT ?a ?b
    WHERE {
        ?a :xxx "xxx"
        OPTIONAL {
            ?b :xxx "XXX"
        }
    }
|]
```

This would require a fairly complete parse of the SPARQL, so would not
be trivial.