All tests now pass
Although the only tests that are present are for RDF generation, and
even these are not complete.

I need to cover the composition cases (e.g. with parenthesis and
application).

I also need to add tests of code generation. To do this I need a new
kind of test suite that compares results to files.