Update TODO and strategize
So I am having a little trouble with my approach to code generating. The
results returned from the SPARQL queries is terribly unwieldy. I get
tables when I want graphs. Then I have to go to great lengths to convert
the table into a form I can use. Further, I am cobbled by a commitment
to minimize the number of SPARQL queries (for performance reasons).

Something needs to change. Abandoning the RDF/SPARQL approach is not an
option. I want to be able to build programs from RDF graphs.

One option is to build specialized Haskell structures from the RDF
(throwing away the ad hoc tuples I currently use). I could write a
typeclass for types derived from Morloc RDF. It would include the
functions for reading from a SPARQL endpoint and for writing to a set of
triples.

A second option is to ignore performance and write SPARQL queries every
time I need data. Retrieving a particular piece of data from SPARQL is
not to difficult. Although, no, it really is. That would require all the
generation process is done in the IO monad. I would have to do error
handling and checking all through the process. And the gymnastics of
working with bad data structures would murder the beauty of my program.

OK, that settles it. I will write data structures.

Now what about the parser? Should I parse directly to the data
structures rather than to RDF? I abandoned that approach a long time
ago. At the time, I was thinking that it was a good idea since it gave
me the flexibility to add new attributes without having to update my
structures. Also, in the past, I had a mess of structures and
intermediate structures and conversions between them, it was a monster.

But now I have a parser that is all tangled up with the details of my
RDF specification. Thus the RDF is biased towards forms that are
convenient to parse, rather than convenient to operate on. This is the
opposite of how it should be.