Wrote SPARQL query to find most manifold info
I can retrieve data like this:

```
----------------------------------------------------------------------------------------------------------
| decname | dargs | callid  | falias         | element | argname | argcallname    | datatype   | dataval |
==========================================================================================================
| "bar"   | "y"   | mlc:101 | "baz"          | rdf:_0  | "y"     |                |            |         |
| "foo"   | "x"   | mlc:110 | "bar"          | rdf:_0  | "x"     |                |            |         |
|         |       | mlc:25  | "len"          | rdf:_0  | "xs"    |                |            |         |
| "roll"  | "d,n" | mlc:79  | "ceiling"      | rdf:_0  |         | "rand_uniform" |            |         |
|         |       | mlc:81  | "rand_uniform" | rdf:_0  | "n"     |                |            |         |
|         |       | mlc:81  | "rand_uniform" | rdf:_1  |         |                | mlc:number | 0.0     |
|         |       | mlc:81  | "rand_uniform" | rdf:_2  | "d"     |                |            |         |
| "baz"   | "z"   | mlc:92  | "ceiling"      | rdf:_0  | "z"     |                |            |         |
----------------------------------------------------------------------------------------------------------
```

This is most of what I need to build the manifolds. Not everything. But
it is a start. Also I will need some functions to trickle values down to
children (specifically `dargs`).

The handling of literal data is also far from complete. Currently what I
return, just the top type, will work fine for data primitives (number,
string, and bool), but not for any of the collections.

It would also be worthwhile to add in the packers and unpackers. Though
I could alternatively add find these independently (especially since the
packer/unpackers are language-specific).

Also I should add in source and export info.

And type. At least if I want to do any type checking. Though the types
are also nice for documentation and usage statements.